requirements.txt
pefile
pycryptodome